refactor: refatora os controladores e seus respectivos testes @Geoziihdev @Lip29bm
